+++ policy/tests/macro_tests.rs
@@ -1,22 +1,35 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
-use policy_macros::{impl_count_tuples, map};
+use policy_macros::map;
 use std::collections::HashMap;
 
-/// Local type that will be implemented by the trait
-pub trait Count {
-    /// Returns the number of items of an implementor
-    fn count(&self) -> usize;
-}
-
-// tuple count trait must be called here
-impl_count_tuples!(16);
+use policyengine::types::{access::Access, Cardinality};
 
 #[test]
-pub fn test_tuple_count() {
-    assert_eq!((1, 2, 3, 4).count(), 4);
-    assert_eq!((1, 2, 3, 4, "string").count(), 5);
-    assert_eq!((1, 2, 3, 4, 232.32, 34, 'a', "other string").count(), 8);
+#[allow(dead_code)]
+fn test_enum_cardinality() {
+    #[derive(Cardinality)]
+    enum LocalA {
+        A,
+        B,
+        C,
+    }
+
+    #[derive(Cardinality)]
+    enum LocalB<T>
+    where
+        T: Clone + Default,
+    {
+        A,
+        Random,
+        SomeValue(usize),
+        StructVariant { name: String },
+        Generic(T),
+    }
+
+    assert_eq!(LocalA::cardinality(), 3);
+    assert_eq!(LocalB::<usize>::cardinality(), 5);
+    assert_eq!(Access::cardinality(), 5);
 }
 
 #[test]
